requirements.txt
pypresence @ git+https://github.com/qwertyquerty/pypresence.git
pyinstaller==6.5.0
PyQtWebEngine
yandex-music
packaging
requests
colorama
pystray
keyring
pywin32
Pillow
asyncio
psutil
PyQt6
PyQt6-WebEngine
aiohttp